refactor where signals are registered
